+++ src/content/layout/builtin_layout.cc
@@ -86,7 +86,10 @@ BuiltinLayout::BuiltinLayout(std::shared_ptr<Content> content)
 #endif
 }
 
-int BuiltinLayout::add(const std::shared_ptr<CdsObject>& obj, const std::pair<int, bool>& parentID, bool useRef)
+int BuiltinLayout::add(
+    const std::shared_ptr<CdsObject>& obj,
+    const std::pair<int, bool>& parentID,
+    bool useRef)
 {
     obj->setParentID(parentID.first);
     if (useRef)
@@ -97,7 +100,12 @@ int BuiltinLayout::add(const std::shared_ptr<CdsObject>& obj, const std::pair<in
     return obj->getID();
 }
 
-int BuiltinLayout::getDir(const std::shared_ptr<CdsObject>& obj, const fs::path& rootPath, const std::string_view& c1, const std::string_view& c2, const std::string& upnpClass)
+int BuiltinLayout::getDir(
+    const std::shared_ptr<CdsObject>& obj,
+    const fs::path& rootPath,
+    const std::string_view& c1,
+    const std::string_view& c2,
+    const std::string& upnpClass)
 {
     fs::path dir;
     auto&& objPath = obj->getLocation();
@@ -131,7 +139,11 @@ int BuiltinLayout::getDir(const std::shared_ptr<CdsObject>& obj, const fs::path&
     return INVALID_OBJECT_ID;
 }
 
-std::vector<int> BuiltinLayout::addVideo(const std::shared_ptr<CdsObject>& obj, const std::shared_ptr<CdsContainer>& parent, const fs::path& rootpath, const std::map<AutoscanMediaMode, std::string>& containerMap)
+std::vector<int> BuiltinLayout::addVideo(
+    const std::shared_ptr<CdsObject>& obj,
+    const std::shared_ptr<CdsContainer>& parent,
+    const fs::path& rootpath,
+    const std::map<AutoscanMediaMode, std::string>& containerMap)
 {
     auto f2i = converterManager->f2i();
     auto blOption = config->getBoxLayoutListOption(ConfigVal::BOXLAYOUT_BOX);
@@ -196,7 +208,11 @@ std::vector<int> BuiltinLayout::addVideo(const std::shared_ptr<CdsObject>& obj, 
     return result;
 }
 
-std::vector<int> BuiltinLayout::addImage(const std::shared_ptr<CdsObject>& obj, const std::shared_ptr<CdsContainer>& parent, const fs::path& rootpath, const std::map<AutoscanMediaMode, std::string>& containerMap)
+std::vector<int> BuiltinLayout::addImage(
+    const std::shared_ptr<CdsObject>& obj,
+    const std::shared_ptr<CdsContainer>& parent,
+    const fs::path& rootpath,
+    const std::map<AutoscanMediaMode, std::string>& containerMap)
 {
     auto f2i = converterManager->f2i();
     auto blOption = config->getBoxLayoutListOption(ConfigVal::BOXLAYOUT_BOX);
@@ -247,7 +263,7 @@ std::vector<int> BuiltinLayout::addImage(const std::shared_ptr<CdsObject>& obj, 
             }
             std::vector<std::shared_ptr<CdsObject>> ct;
             ct.push_back(containerAt(BoxKeys::imageRoot));
-            ct.push_back(containerAt(BoxKeys::imageAllYears));
+            ct.push_back(containerAt(BoxKeys::imageAllDates));
             ct.push_back(std::make_shared<CdsContainer>(date));
             id = content->addContainerTree(ct, obj);
             result.push_back(add(obj, id));
@@ -260,7 +276,11 @@ std::vector<int> BuiltinLayout::addImage(const std::shared_ptr<CdsObject>& obj, 
     return result;
 }
 
-std::vector<int> BuiltinLayout::addAudio(const std::shared_ptr<CdsObject>& obj, const std::shared_ptr<CdsContainer>& parent, const fs::path& rootpath, const std::map<AutoscanMediaMode, std::string>& containerMap)
+std::vector<int> BuiltinLayout::addAudio(
+    const std::shared_ptr<CdsObject>& obj,
+    const std::shared_ptr<CdsContainer>& parent,
+    const fs::path& rootpath,
+    const std::map<AutoscanMediaMode, std::string>& containerMap)
 {
     auto f2i = converterManager->f2i();
     auto blOption = config->getBoxLayoutListOption(ConfigVal::BOXLAYOUT_BOX);
@@ -467,7 +487,11 @@ std::vector<int> BuiltinLayout::addAudio(const std::shared_ptr<CdsObject>& obj, 
 }
 
 #ifdef ONLINE_SERVICES
-std::vector<int> BuiltinLayout::addOnlineItem(const std::shared_ptr<CdsObject>& obj, OnlineServiceType serviceType, const fs::path& rootpath, const std::map<AutoscanMediaMode, std::string>& containerMap)
+std::vector<int> BuiltinLayout::addOnlineItem(
+    const std::shared_ptr<CdsObject>& obj,
+    OnlineServiceType serviceType,
+    const fs::path& rootpath,
+    const std::map<AutoscanMediaMode, std::string>& containerMap)
 {
     switch (serviceType) {
     case OnlineServiceType::Max: